// src/shared/types/product.type.ts
import mongoose from 'mongoose';
import { ProductStatus } from '../enums/product-status.enum';
import { ProductType } from '../enums/product-type.enum';
import { Coupon } from './coupon.type';
import { Review } from './review.type';
import { Question } from './question.type';

export type Product = {
  title: string;
  pageTitle?: string;
  description: string;
  bullet_points: string;
  price: number;
  salePrice?: number;
  percentageDiscount?: number;
  showDiscount?: boolean;
  image_url: string;
  images: string[];
  rating: number;
  coupon?: Coupon;
  status?: ProductStatus;
  asin?: string;
  image?: string;
  file?: string;
  fileText?: string;
  productType?: ProductType;
  reviews: Review[];
  video?: string;
  link?: string;
  questions?: Question[];
  pageColors?: Record<any, string>;
  fileDownloadWithEmail?: boolean;
  pageViews?: number;
  takenCoupons?: number;
  linkClicks?: number;
  filesDownloaded?: number;
  conversationRate?: mongoose.Types.Decimal128;
  unlimitedCoupons?: boolean;
  takenCouponByUser?: boolean;
  created_at?: Date;
};